Check jason file exist

+++ steps/Run_IBM_ASR.py
@@ -9,6 +9,7 @@ import json
 import pandas as pd
 import numpy as np
 from collections import defaultdict
+from os.path import join, isfile
 
 #TODO replace with argparse
 dir = sys.argv[1]
@@ -26,21 +27,18 @@ if not os.path.isdir(tmp_dir):
 #TODO check wav file is exist
 
 wav_file = os.path.join(dir,'{}_{}.wav'.format(childID, taskID))
-json_file = os.path.join(tmp_dir,'{}_{}_ibm.json'.format(childID, taskID))
+json_file = os.path.join(dir,'{}_{}_ibm.json'.format(childID, taskID))
 txtgrid_file = os.path.join(dir,'{}_{}_ibm.TextGrid'.format(childID, taskID))
 
 #Get duration
 dur_in_secs = librosa.get_duration(filename=wav_file)
-
-#TODO check if json exist, enable force rerun
-results = ibm_stt.stt_audio_file(wav_file,model_str='BroadbandModel')
-
-with open(json_file,'w') as fjson:
-    json.dump(results, fjson)
-
-#with open(json_file) as fjson:
-#    results = json.load(fjson)
-    
+if isfile(json_file):
+    with open(json_file) as fjson:
+        results = json.load(fjson)
+else:
+    results = ibm_stt.stt_audio_file(wav_file,model_str='BroadbandModel')
+    with open(json_file,'w') as fjson:
+        json.dump(results, fjson)
 
 
 df_spkrs = pd.DataFrame.from_records(results[0]['speaker_labels'])
